# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5eb64906bbf7210dd3bd2f89
  title: WeddingRing
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  "/cities/{id}":
    delete:
      x-skaffolder-id: 5ec0331da940665206b01192
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5ec02f21a940665206b0114c
      x-skaffolder-resource: Cities
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Cities
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5ec0331da940665206b0118e
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5ec02f21a940665206b0114c
      x-skaffolder-resource: Cities
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Cities
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cities"
    post:
      x-skaffolder-id: 5ec0331ca940665206b0118c
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5ec02f21a940665206b0114c
      x-skaffolder-resource: Cities
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Cities
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cities"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Cities"
...
  "/maritalstatus/{id}":
    delete:
      x-skaffolder-id: 5ec03324a940665206b0119c
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5ec03061a940665206b0115c
      x-skaffolder-resource: Maritalstatus
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Maritalstatus
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5ec03324a940665206b01197
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5ec03061a940665206b0115c
      x-skaffolder-resource: Maritalstatus
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Maritalstatus
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Maritalstatus"
    post:
      x-skaffolder-id: 5ec03324a940665206b01199
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5ec03061a940665206b0115c
      x-skaffolder-resource: Maritalstatus
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Maritalstatus
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Maritalstatus"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Maritalstatus"
...
  "/states/{id}":
    delete:
      x-skaffolder-id: 5ec03314a940665206b01189
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5ebfef21a940665206b00fed
      x-skaffolder-resource: States
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - States
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5ec03313a940665206b01182
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5ebfef21a940665206b00fed
      x-skaffolder-resource: States
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - States
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/States"
    post:
      x-skaffolder-id: 5ec03313a940665206b01184
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5ebfef21a940665206b00fed
      x-skaffolder-resource: States
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - States
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/States"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/States"
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Cities:
      x-skaffolder-id: 5ec02f21a940665206b0114c
      x-skaffolder-id-db: 5eb64906bbf7210dd3bd2f8a
      x-skaffolder-id-entity: 5ec02f20a940665206b01149
      x-skaffolder-url: /cities
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        CityName:
          type: string
          x-skaffolder-id-attr: 5ec02f20a940665206b0114a
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations:
        CityName:
          x-skaffolder-id: 5ec02f20a940665206b0114b
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 5ec02f20a940665206b01149
          x-skaffolder-ent2: 5eb64c32bbf7210dd3bd300e
      required:
        - _id
        - CityName
    Maritalstatus:
      x-skaffolder-id: 5ec03061a940665206b0115c
      x-skaffolder-id-db: 5eb64906bbf7210dd3bd2f8a
      x-skaffolder-id-entity: 5ec03061a940665206b01159
      x-skaffolder-url: /maritalstatus
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        Maritalstatus:
          type: string
          x-skaffolder-id-attr: 5ec03061a940665206b0115a
          x-skaffolder-type: String
      x-skaffolder-relations:
        MartialStatus:
          x-skaffolder-id: 5ec03061a940665206b0115b
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 5ec03061a940665206b01159
          x-skaffolder-ent2: 5eb64c32bbf7210dd3bd300e
      required:
        - _id
    States:
      x-skaffolder-id: 5ebfef21a940665206b00fed
      x-skaffolder-id-db: 5eb64906bbf7210dd3bd2f8a
      x-skaffolder-id-entity: 5ebfef21a940665206b00feb
      x-skaffolder-url: /states
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        Name:
          type: string
          x-skaffolder-id-attr: 5ebfef21a940665206b00fec
          x-skaffolder-type: String
          x-skaffolder-unique: true
      x-skaffolder-relations:
        currentState:
          x-skaffolder-id: 5ec02ef9a940665206b01147
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 5ebfef21a940665206b00feb
          x-skaffolder-ent2: 5eb64c32bbf7210dd3bd300e
      required:
        - _id
...